Move segyio/_field.py -> segyio/field.py
