+++ python/segyio/field.py
